+++ mainservice/src/main/java/sustech/ooad/mainservice/mapper/AuthUserRepository.java
@@ -1,8 +1,8 @@
 package sustech.ooad.mainservice.mapper;
 
+import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sustech.ooad.mainservice.model.AuthUser;
 
-public interface AuthUserRepository extends JpaRepository<AuthUser, Long> {
-
+public interface AuthUserRepository extends JpaRepository<AuthUser, BigDecimal> {
 }
